movie_explorer/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
streamlit
requests
python-dotenv
langchain
langchain-groq
